Rename: 테스트 클래스 이름 변경
